requirements/requirements.txt
Django==1.4
MySQL-python==1.2.4
South==0.7.6
django-extensions==1.1.1
django-debug-toolbar==0.9.4
django-admin-honeypot==0.2.5

django-braces==1.0.0

django-crispy-forms==1.2.5
django-floppyforms==1.1

Pillow
django-userena==1.0.1

pytz

django-social-auth==0.7.22

django-jsonfield
django-activity-stream

django-tinymce==1.5.1

python-memcached==1.31

coverage==3.6
django-discover-runner==0.4
django-cache-panel
